fix: correct parseInt radix parameters and use parseFloat for dimensions
- Fix parseInt calls using wrong radix (3, 4) instead of 10 in sot323,
  sot23w, sot343 pin count parsing
- Fix parseInt missing radix in to220, to220f, bga
- Replace parseInt with parseFloat for height parameters in sot23,
  sot23w, sot323 silkscreen positioning to preserve decimal precision

These bugs were latent (current inputs happened to work) but would cause
incorrect behavior with certain parameter values.

+++ src/fn/sot23.ts
@@ -102,7 +102,7 @@ export const sot23_3 = (parameters: z.infer<typeof sot23_def>) => {
   }
   const silkscreenRefText: SilkscreenRef = silkscreenRef(
     0,
-    Number.parseInt(parameters.h),
+    Number.parseFloat(parameters.h),
     0.3,
   )
   return [...pads, silkscreenRefText as AnyCircuitElement]
